azuredevops_agent/pyproject.toml
[project]
name = "azuredevops-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-devops>=7.1.0b4",
    "chainlit>=1.3.0",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.0.0",
]
